+++ tests/courseformatrenderer_test.php
@@ -19,13 +19,13 @@
  *
  * @package    course/format
  * @subpackage topcoll
- * @version    See the value of '$plugin->version' in below.
+ * @version    See the value of '$plugin->version' in version.php.
  * @copyright  &copy; 2015-onwards G J Barnard in respect to modifications of standard topics format.
- * @author     G J Barnard - gjbarnard at gmail dot com, {@link http://about.me/gjbarnard} and
- *                           {@link http://moodle.org/user/profile.php?id=442195}
+ * @author     G J Barnard - {@link http://moodle.org/user/profile.php?id=442195}
  * @link       http://docs.moodle.org/en/Collapsed_Topics_course_format
  * @license    http://www.gnu.org/copyleft/gpl.html GNU Public License
  */
+defined('MOODLE_INTERNAL') || die();
 
 /**
  * Renderer unit tests for the Collapsed Topics course format.
@@ -127,18 +127,21 @@ class format_topcoll_courseformatrenderer_testcase extends advanced_testcase {
     }
 
     public function test_section_right_content() {
+        global $CFG;
+
         // With defaults unchanged.
         $section = $this->courseformat->get_section(1);
         $onsectionpage = false;
         $theclass = self::call_method($this->outputus, 'section_right_content',
             array($section, $this->course, $onsectionpage));
-        $thevalue = '<span class="cps_centre">Topic<br />1</span>';
+        $thevalue = '<a title="View only &#039;Topic 1&#039;" class="cps_centre" ';
+        $thevalue .= 'href="'.$CFG->wwwroot.'/course/view.php?id='.$this->course->id.'&amp;section=1">Topic<br />1</a>';
         $this->assertEquals($thevalue, $theclass);
 
         $onsectionpage = true;
         $theclass = self::call_method($this->outputus, 'section_right_content',
             array($section, $this->course, $onsectionpage));
-        $thevalue = '<span class="cps_centre">Topic<br />1</span>';
+        $thevalue = '';
         $this->assertEquals($thevalue, $theclass);
     }
 
@@ -153,7 +156,7 @@ class format_topcoll_courseformatrenderer_testcase extends advanced_testcase {
         $onsectionpage = true;
         $theclass = self::call_method($this->outputus, 'section_left_content',
             array($section, $this->course, $onsectionpage));
-        $thevalue = '<span class="cps_centre">1</span>';
+        $thevalue = '';
         $this->assertEquals($thevalue, $theclass);
     }
 
@@ -166,12 +169,15 @@ class format_topcoll_courseformatrenderer_testcase extends advanced_testcase {
         $thevalue = '<li id="section-1" class="section main section-summary clearfix" role="region" aria-label="';
         $thevalue .= 'Section 1"><div class="left side"></div><div class="right side"></div><div class="content">';
         $thevalue .= '<h3 class="section-title"><a href="'.$CFG->wwwroot.'/course/view.php?id=';
-        $thevalue .= $this->course->id.'#section-1" class="">Section 1</a></h3><div class="summarytext"></div></div></li>';
+        $thevalue .= $this->course->id.'#section-1" class="">Section 1</a></h3><div class="summarytext"></div>';
+        $thevalue .= '<div class="section_availability"></div></div></li>';
 
         $this->assertEquals($thevalue, $theclass);
     }
 
     public function test_section_header() {
+        global $CFG;
+
         $section = $this->courseformat->get_section(1);
         $section->toggle = false;
 
@@ -180,56 +186,67 @@ class format_topcoll_courseformatrenderer_testcase extends advanced_testcase {
             array($section, $this->course, $onsectionpage));
         $thevalue = '<li id="section-1" class="section main clearfix span12" role="region" aria-label="Section 1">';
         $thevalue .= '<div class="left side"><span class="cps_centre">1</span></div><div class="right side">';
-        $thevalue .= '<span class="cps_centre">Topic<br />1</span></div><div class="content">';
-        $thevalue .= '<div class="sectionhead toggle toggle-arrow" id="toggle-1">';
-        $thevalue .= '<span class="toggle_closed the_toggle tc-medium" ';
-        $thevalue .= 'role="button" aria-pressed="false">';
-        $thevalue .= '<h3 class="sectionname">Section 1<div class="cttoggle"> - Toggle</div></h3></span></div>';
+        $thevalue .= '<a title="View only &#039;Topic 1&#039;" class="cps_centre" ';
+        $thevalue .= 'href="'.$CFG->wwwroot.'/course/view.php?id='.$this->course->id.'&amp;section=1">Topic<br />1</a>';
+        $thevalue .= '</div><div class="content"><div class="sectionhead toggle toggle-arrow" id="toggle-1">';
+        $thevalue .= '<span class="toggle_closed the_toggle tc-medium" role="button" aria-pressed="false">';
+        $thevalue .= '<h3 class="sectionname">Section 1<div class="cttoggle"> - Toggle</div></h3>';
+        $thevalue .= '<div class="section_availability"></div></span></div>';
         $thevalue .= '<div class="sectionbody toggledsection" id="toggledsection-1">';
         $this->assertEquals($thevalue, $theclass);
-
         $onsectionpage = true;
         $theclass = self::call_method($this->outputus, 'section_header',
             array($section, $this->course, $onsectionpage));
         $thevalue = '<li id="section-1" class="section main clearfix span12" role="region" aria-label="Section 1">';
-        $thevalue .= '<div class="left side"><span class="cps_centre">1</span></div><div class="right side"><span ';
-        $thevalue .= 'class="cps_centre">Topic<br />1</span></div><div class="content"><div class="summary"></div>';
+        $thevalue .= '<div class="left side"></div><div class="right side"></div><div class="content">';
+        $thevalue .= '<div class="section_availability"></div><div class="summary"></div>';
 
         $this->assertEquals($thevalue, $theclass);
     }
 
     public function test_stealth_section_header() {
+        global $CFG;
+
         $theclass = self::call_method($this->outputus, 'stealth_section_header',
             array(1));
-        $thevalue = '<li id="section-1" class="section main clearfix orphaned hidden span12" role="region" aria-label=';
-        $thevalue .= '"Section 1"><div class="left side"></div><div class="right side"><span class="cps_centre">Topic';
-        $thevalue .= '<br />1</span></div><div class="content"><h3 class="sectionname">Orphaned activities (section 1)</h3>';
+        $thevalue = '<li id="section-1" class="section main clearfix orphaned hidden span12" role="region" ';
+        $thevalue .= 'aria-label="Section 1"><div class="left side"></div><div class="right side">';
+        $thevalue .= '<a title="View only &#039;Topic 1&#039;" class="cps_centre" ';
+        $thevalue .= 'href="'.$CFG->wwwroot.'/course/view.php?id='.$this->course->id.'&amp;section=1">Topic<br />1</a>';
+        $thevalue .= '</div><div class="content"><h3 class="sectionname">Orphaned activities (section 1)</h3>';
 
         $this->assertEquals($thevalue, $theclass);
     }
 
     public function test_section_hidden() {
+        global $CFG;
+
         $section = $this->courseformat->get_section(1);
         $theclass = self::call_method($this->outputus, 'section_hidden',
             array($section, null));
         $thevalue = '<li id="section-1" class="section main clearfix hidden span12" role="region" aria-label="Section 1">';
-        $thevalue .= '<div class="left side"><span class="cps_centre">1</span></div><div class="right side"><span class="';
-        $thevalue .= 'cps_centre">Topic<br />1</span></div><div class="content sectionhidden"><h3 class="section-title">';
-        $thevalue .= 'Not available</h3></div></li>';
+        $thevalue .= '<div class="left side"><span class="cps_centre">1</span></div><div class="right side">';
+        $thevalue .= '<a title="View only &#039;Topic 1&#039;" class="cps_centre" ';
+        $thevalue .= 'href="'.$CFG->wwwroot.'/course/view.php?id='.$this->course->id.'&amp;section=1">Topic<br />1</a></div>';
+        $thevalue .= '<div class="content sectionhidden"><h3 class="section-title">Not available</h3></div></li>';
 
         $this->assertEquals($thevalue, $theclass);
     }
 
     public function test_print_multiple_section_page() {
+        global $CFG;
+
         self::call_method($this->outputus, 'print_multiple_section_page',
             array($this->course, null, null, null, null, null));
         $theoutput = '<h2 class="accesshide">Section</h2><ul class="ctopics"></ul><ul class="ctopics topics row-fluid">';
         $theoutput .= '<li id="section-1" class="section main clearfix span12" role="region" aria-label="Section 1">';
         $theoutput .= '<div class="left side"><span class="cps_centre">1</span></div><div class="right side">';
-        $theoutput .= '<span class="cps_centre">Topic<br />1</span></div><div class="content"><div class="sectionhead ';
-        $theoutput .= 'toggle toggle-arrow" id="toggle-1"><span class="toggle_closed the_toggle tc-medium" role="button" ';
-        $theoutput .= 'aria-pressed="false">';
-        $theoutput .= '<h3 class="sectionname">Section 1<div class="cttoggle"> - Toggle</div></h3></span></div>';
+        $theoutput .= '<a title="View only &#039;Topic 1&#039;" class="cps_centre" ';
+        $theoutput .= 'href="'.$CFG->wwwroot.'/course/view.php?id='.$this->course->id.'&amp;section=1">Topic<br />1</a>';
+        $theoutput .= '</div><div class="content"><div class="sectionhead toggle toggle-arrow" id="toggle-1">';
+        $theoutput .= '<span class="toggle_closed the_toggle tc-medium" role="button" aria-pressed="false">';
+        $theoutput .= '<h3 class="sectionname">Section 1<div class="cttoggle"> - Toggle</div></h3>';
+        $theoutput .= '<div class="section_availability"></div></span></div>';
         $theoutput .= '<div class="sectionbody toggledsection" id="toggledsection-1"><ul class="section img-text">';
         $theoutput .= '</ul></div></div></li></ul>';
         $this->expectOutputString($theoutput);
@@ -239,10 +256,10 @@ class format_topcoll_courseformatrenderer_testcase extends advanced_testcase {
         global $CFG;
 
         $theclass = self::call_method($this->outputus, 'toggle_all', array());
-        $thevalue = '<li class="tcsection main clearfix" id="toggle-all"><div class="left side"><img width="1" height="1" ';
-        $thevalue .= 'class="spacer" alt="" src="'.$CFG->wwwroot.'/theme/image.php/_s/clean/core/1/spacer"/>';
-        $thevalue .= '</div><div class="right side"><img width="1" height="1" class="spacer" alt="" src="';
-        $thevalue .= $CFG->wwwroot.'/theme/image.php/_s/clean/core/1/spacer"/></div><div class="content">';
+        $thevalue = '<li class="tcsection main clearfix" id="toggle-all"><div class="left side"><img class="icon spacer" ';
+        $thevalue .= 'width="1" height="1" alt="" src="'.$CFG->wwwroot.'/theme/image.php/_s/clean/core/1/spacer" />';
+        $thevalue .= '</div><div class="right side"><img class="icon spacer" width="1" height="1" alt="" src="';
+        $thevalue .= $CFG->wwwroot.'/theme/image.php/_s/clean/core/1/spacer" /></div><div class="content">';
         $thevalue .= '<div class="sectionbody toggle-arrow-hover toggle-arrow"><h4><span class="on tc-medium" id="';
         $thevalue .= 'toggles-all-opened" role="button">Open all</span><span class="off tc-medium" id="toggles-all-closed" ';
         $thevalue .= 'role="button">Close all</span></h4></div></div></li>';
@@ -254,10 +271,10 @@ class format_topcoll_courseformatrenderer_testcase extends advanced_testcase {
 
         $theclass = self::call_method($this->outputus, 'display_instructions', array());
         $thevalue = '<li class="tcsection main clearfix" id="topcoll-display-instructions"><div class="left side">';
-        $thevalue .= '<img width="1" height="1" class="spacer" alt="" src="';
-        $thevalue .= $CFG->wwwroot.'/theme/image.php/_s/clean/core/1/spacer"/></div><div class="right side">';
-        $thevalue .= '<img width="1" height="1" class="spacer" alt="" src="';
-        $thevalue .= $CFG->wwwroot.'/theme/image.php/_s/clean/core/1/spacer"/></div><div class="content">';
+        $thevalue .= '<img class="icon spacer" width="1" height="1" alt="" src="';
+        $thevalue .= $CFG->wwwroot.'/theme/image.php/_s/clean/core/1/spacer" /></div><div class="right side">';
+        $thevalue .= '<img class="icon spacer" width="1" height="1" alt="" src="';
+        $thevalue .= $CFG->wwwroot.'/theme/image.php/_s/clean/core/1/spacer" /></div><div class="content">';
         $thevalue .= '<div class="sectionbody"><p class="topcoll-display-instructions">Instructions: Clicking on the section ';
         $thevalue .= 'name will show / hide the section.</p></div></div></li>';